Module pallet_collective::pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

Re-exports

pub use __tt_default_parts_6 as tt_default_parts;

Structs

GenesisConfig

Can be used to configure the genesis state of this pallet.

Instance1

Instance1 to be used for instantiable pallet define with pallet macro.

Instance2

Instance2 to be used for instantiable pallet define with pallet macro.

Instance3

Instance3 to be used for instantiable pallet define with pallet macro.

Instance4

Instance4 to be used for instantiable pallet define with pallet macro.

Instance5

Instance5 to be used for instantiable pallet define with pallet macro.

Instance6

Instance6 to be used for instantiable pallet define with pallet macro.

Instance7

Instance7 to be used for instantiable pallet define with pallet macro.

Instance8

Instance8 to be used for instantiable pallet define with pallet macro.

Instance9

Instance9 to be used for instantiable pallet define with pallet macro.

Instance10

Instance10 to be used for instantiable pallet define with pallet macro.

Instance11

Instance11 to be used for instantiable pallet define with pallet macro.

Instance12

Instance12 to be used for instantiable pallet define with pallet macro.

Instance13

Instance13 to be used for instantiable pallet define with pallet macro.

Instance14

Instance14 to be used for instantiable pallet define with pallet macro.

Instance15

Instance15 to be used for instantiable pallet define with pallet macro.

Instance16

Instance16 to be used for instantiable pallet define with pallet macro.

Pallet

The pallet implementing the on-chain logic.

_GeneratedPrefixForStorageMembers

_GeneratedPrefixForStoragePrime

_GeneratedPrefixForStorageProposalCount

_GeneratedPrefixForStorageProposalOf

_GeneratedPrefixForStorageProposals

_GeneratedPrefixForStorageVoting

Enums

Call

Contains one variant per dispatchable that can be called by an extrinsic.

Error

Custom dispatch errors of this pallet.

Event

The event emitted by this pallet.

Traits

Config

Configuration trait of this pallet.

Type Definitions

Members

The current members of the collective. This is stored sorted (just by value).

ModuleDeprecated

Type alias to Pallet, to be used by construct_runtime.

Origin

Origin for the collective pallet.

Prime

The prime member that helps determine the default vote behavior in case of absentations.

ProposalCount

Proposals so far.

ProposalOf

Actual proposal for a given hash, if it’s current.

Proposals

The hashes of the active proposals.

RawEventDeprecated

Old name generated by decl_event.

Voting

Votes on a given proposal, if it is ongoing.