pallet_collective::pallet::dispatchables

Function execute

pub fn execute<T: Config<I>, I: 'static>(
    proposal: Box<<T as Config<I>>::Proposal>,
    length_bound: u32,
)

Dispatch a proposal from a member using the Member origin.

Origin must be a member of the collective.

Complexity:

O(B + M + P) where:

B is proposal size in bytes (length-fee-bounded)

M members-count (code-bounded)

P complexity of dispatching proposal

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::execute.