Function pallet_collective::pallet::dispatchables::propose

pub fn propose<T: Config<I>, I: 'static>(
    threshold: MemberCount,
    proposal: Box<<T as Config<I>>::Proposal>,
    length_bound: u32,
)

Add a new proposal to either be voted on or executed directly.

Requires the sender to be member.

threshold determines whether proposal is executed directly (threshold < 2) or put up for voting.

Complexity

O(B + M + P1) or O(B + M + P2) where:

B is proposal size in bytes (length-fee-bounded)

M is members-count (code- and governance-bounded)

branching is influenced by threshold where:

P1 is proposal execution complexity (threshold < 2)

P2 is proposals-count (code-bounded) (threshold >= 2)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::propose.