Trait pallet_collective::pallet::Config

pub trait Config<I: 'static = ()>: Config {
    type RuntimeOrigin: From<RawOrigin<Self::AccountId, I>>;
    type Proposal: Parameter + Dispatchable<RuntimeOrigin = <Self as Config<I>>::RuntimeOrigin, PostInfo = PostDispatchInfo> + From<Call<Self>> + GetDispatchInfo;
    type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>;
    type MotionDuration: Get<Self::BlockNumber>;
    type MaxProposals: Get<ProposalIndex>;
    type MaxMembers: Get<MemberCount>;
    type DefaultVote: DefaultVote;
    type WeightInfo: WeightInfo;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeOrigin: From<RawOrigin<Self::AccountId, I>>

The runtime origin type.

type Proposal: Parameter + Dispatchable<RuntimeOrigin = <Self as Config<I>>::RuntimeOrigin, PostInfo = PostDispatchInfo> + From<Call<Self>> + GetDispatchInfo

The runtime call dispatch type.

type RuntimeEvent: From<Event<Self, I>> + IsType<<Self as Config>::RuntimeEvent>

The runtime event type.

type MotionDuration: Get<Self::BlockNumber>

The time-out for council motions.

type MaxProposals: Get<ProposalIndex>

Maximum number of proposals allowed to be active in parallel.

type MaxMembers: Get<MemberCount>

The maximum number of members supported by the pallet. Used for weight estimation.

NOTE:

Benchmarks will need to be re-run and weights adjusted if this changes.

This pallet assumes that dependents keep to the limit without enforcing it.

type DefaultVote: DefaultVote

Default vote strategy of this collective.

type WeightInfo: WeightInfo

Weight information for extrinsics in this pallet.