pallet_collective::pallet

Type Alias Origin

pub type Origin<T, I = ()> = RawOrigin<<T as Config>::AccountId, I>;

Origin for the collective pallet.

Aliased Type

enum Origin<T, I = ()> {
    Members(u32, u32),
    Member(<T as Config>::AccountId),
    _Phantom(PhantomData<I>),
}

Variants

Members(u32, u32)

It has been condoned by a given number of members of the collective from a given total.

Member(<T as Config>::AccountId)

It has been condoned by a single member of the collective.

_Phantom(PhantomData<I>)

Dummy to manage the fact we have instancing.