Struct pallet_collective::Module

pub struct Module<T: Config<I>, I: Instance = DefaultInstance>(_);

Implementations

impl<T: Config<I> + 'static, I: Instance> Module<T, I>

pub fn proposals() -> BoundedVec<T::Hash, T::MaxProposals>

The hashes of the active proposals.

pub fn proposal_of<K: EncodeLike<T::Hash>>(
key: K
) -> Option<<T as Config<I>>::Proposal>

Actual proposal for a given hash, if it’s current.

pub fn voting<K: EncodeLike<T::Hash>>(
key: K
) -> Option<Votes<T::AccountId, T::BlockNumber>>

Votes on a given proposal, if it is ongoing.

pub fn proposal_count() -> u32

Proposals so far.

pub fn members() -> Vec<T::AccountId>

The current members of the collective. This is stored sorted (just by value).

pub fn prime() -> Option<T::AccountId>

The prime member that helps determine the default vote behavior in case of absentations.

impl<T: Config<I>, I: Instance> Module<T, I>

pub fn is_member(who: &T::AccountId) -> bool

Check whether who is a member of the collective.