pallet_collective::pallet::dispatchables

Function set_members

pub fn set_members<T: Config<I>, I: 'static>(
    new_members: Vec<T::AccountId>,
    prime: Option<T::AccountId>,
    old_count: MemberCount,
)

Set the collective’s membership.

new_members: The new member list. Be nice to the chain and provide it sorted.

prime: The prime member whose vote sets the default.

old_count: The upper bound for the previous number of members in storage. Used for weight estimation.

The dispatch of this call must be SetMembersOrigin.

NOTE: Does not enforce the expected MaxMembers limit on the amount of members, but the weight estimations rely on it to estimate dispatchable weight.

WARNING:

The pallet-collective can also be managed by logic outside of the pallet through the implementation of the trait [ChangeMembers]. Any call to set_members must be careful that the member set doesn’t get out of sync with other logic managing the member set.

Complexity:

O(MP + N) where:

M old-members-count (code- and governance-bounded)

N new-members-count (code- and governance-bounded)

P proposals-count (code-bounded)

Warning: Doc-Only

This function is an automatically generated, and is doc-only, uncallable stub. See the real version in Pallet::set_members.